feat: 비대면 토론 엔티티 생성 및 객체 분리
